% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_function.R
\name{clean.cdr3}
\alias{clean.cdr3}
\title{We check the first aa is compatible with the V gene and the last two with the J gene.
Cases where one aa is missing due to a different definition of CDR3 are corrected when the first/last 3 aa are compatible with the V/J gene
Cases where the V/J genes are not given are only kept if starting with C and ending with F/W
Cases of CDR3 incompatible with V/J are put to "".}
\usage{
clean.cdr3(v, j, cdr3, sp, chain)
}
\description{
We check the first aa is compatible with the V gene and the last two with the J gene.
Cases where one aa is missing due to a different definition of CDR3 are corrected when the first/last 3 aa are compatible with the V/J gene
Cases where the V/J genes are not given are only kept if starting with C and ending with F/W
Cases of CDR3 incompatible with V/J are put to "".
}
