% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean.name.allele}
\alias{clean.name.allele}
\title{Take the gene + allele.
If allele is empty, try to correct the gene if needed, and return only the gene.
If allele is not empty, try to correct the gene\*allele.
If the gene can be corrected, but the gene\*allele does not exist, return gene\*default.allele
If the gene cannot be corrected, return NA.}
\usage{
clean.name.allele(
  gene = gene,
  allele = allele,
  species = "HomoSapiens",
  seq.protocol = "Default",
  use.allele = F
)
}
\description{
Take the gene + allele.
If allele is empty, try to correct the gene if needed, and return only the gene.
If allele is not empty, try to correct the gene\*allele.
If the gene can be corrected, but the gene\*allele does not exist, return gene\*default.allele
If the gene cannot be corrected, return NA.
}
