% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRviz-package.R
\docType{package}
\name{MixTCRviz.package}
\alias{MixTCRviz-package}
\alias{MixTCRviz.package}
\title{MixTCRviz: Visualize important properties of a set of TCRs}
\description{
MixTCRviz builds TCR binding motifs for a set of TCR.
}
\author{
\strong{Maintainer}: David Gfeller \email{david.gfeller@unil.ch} (\href{https://orcid.org/0000-0002-3952-0930}{ORCID})

Other contributors:
\itemize{
  \item Julien Racle \email{julien.racle@unil.ch} (\href{https://orcid.org/0000-0002-0100-0323}{ORCID}) [contributor]
  \item Daniel Tadros \email{daniel.tadros@unil.ch} [contributor]
}

}
\keyword{internal}
