% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRviz.R
\name{MixTCRviz}
\alias{MixTCRviz}
\title{MixTCRviz: give some title and define the function.}
\usage{
MixTCRviz(
  input1,
  output.path,
  input2 = "",
  baseline.file = "",
  use.allele = 0,
  correct.gene.names = 1,
  use.mouse.strain = 0,
  renormVJ = 1,
  N.min = 10,
  plot = 1,
  plot.cdr12.motif = 0,
  plot.oneline = 0,
  plot.logo.length = 0,
  chain.list.output = "AB",
  input1.name = "Epitope specific",
  input2.name = "",
  output.format = "pdf"
)
}
\arguments{
\item{input1}{csv files or data.frame with the input TCRs. Columns should
ideally include: TRAV, TRAJ, cdr3_TRA, TRBV, TRBJ, cdr3_TRB, model, species
\itemize{
\item The "TRAV", "TRAJ", "TRBV", "TRBJ" should follow the IMGT nomenclature, with or without allele (see below for potential name correction).
If a column is missing, empty values will be used.
\item The "cdr3_TRA" and "cdr3_TRB" columns should provide CDR3A/CDR3B sequences, following the standard definition (e.g., CAVNSDGQKLLF).
If a column is missing, empty values will be used.
\item The "model" column typically describes the epitopes/experiments/classes/... Each model will be treated independently in MixTCRviz.
If missing, all TCRs will be considered as coming from the same model
(Model_default).
\item The "species" column indicates the species of the TCR (e.g., NOT the source organism of the epitope). It should be "HomoSapiens" or "MusMusculus"
If not provided, "HomoSapiens" is taken as default.
}}

\item{output.path}{name of the output directory (if not already existing, it
will be created).}

\item{input2}{(default=""): csv file or data.frame containing the "reference"
TCRs to be used in comparisons. Same format as input1 If provided, the
comparisons is performed with this reference, and not the baseline
repertoire.}

\item{baseline.file}{(default="") .rds file containing information about
baseline repertoire. If empty, the default baseline repertoires are used.}

\item{use.allele}{(default=0)
\itemize{
\item 0: All V/J alleles are merged at the gene level (recommended).
\item 1: Alleles are kept, including mouse TRAV genes from different strains.
Currently, use.allele is fixed to 0 for mouse TCRs.
}}

\item{correct.gene.names}{(default=1)
\itemize{
\item 0: Do not attempt to correct V/J gene names. Put to NA genes not in IMGT.
\item 1: Attempt to correct V/J gene names not in IMGT based on internal map. Put to NA genes that could not be corrected
}}

\item{use.mouse.strain}{(default=0)
\itemize{
\item 0: Merge the different TRAV genes corresponding to different mouse strains (e.g., TRAV10D, TRAVN10 and TRAV10 become all TRAV10).
This is recommended since differences between input TCRs and baseline TCRs
may result from the use of different mouse strains, and not be related to
specificity in input TCRs. In addition different TCR reconstruction tools
use different approaches to call these genes, which can create artificial
enrichment in one of them versus the baseline.
\item 1: The different TRAV segments (e.g., TRAV10D, TRAVN10 and TRAV10) are treated separately.
}}

\item{renormVJ}{(default=1)
\itemize{
\item 0: Use CDR3 length distribution and motif from the full repertoire (this is the only option when input2 is provided for comparison with a user-given reference)
\item 1: When comparing to the baseline TCR repertoire,
build CDR3 length distributions and CDR3 motifs corresponding to the V-J
usage observed in the input TCRs. In the plots the mark " | VJ" is used to
indicate that CDR3 length distributions and motifs correspond to those
expected with the V-J usage in the input TCRs.
}}

\item{N.min}{(default=10) Minimum number of TCR (i.e., V-J-CDR3) for at least
one chain.}

\item{plot}{(default=1)
\itemize{
\item 0: only create .rds object with the statistics
\item 1: Plot the data in output.path/plots/ and create .rds object with the statistics for each model in output.path/stats/.
}}

\item{plot.cdr12.motif}{(default=0)
\itemize{
\item 0: Only show motifs for CDR3.
\item 1: Include sequence motifs for CDR1 and CDR2.
}}

\item{plot.oneline}{(default=0)
\itemize{
\item 0: Show the data on two lines (better for clarity).
\item 1: Show all plots in a single line (can be useful to compare different models).
}}

\item{plot.logo.length}{(default=0)
\itemize{
\item 0: Show only the CDR3 motifs for the most frequent CDR3 length.
\item 1: Show in a separate plot the V usage, J usage and CDR3 motifs for all CDR3 length, for both alpha and beta chains.
}}

\item{chain.list.output}{(default="AB")
\itemize{
\item A: Only the alpha chain is plotted in output;
\item B: only the beta chain is plotted in output;
\item AB both chains are plotted in output
}}

\item{input1.name}{(default="Epitope Specific"): Provide a generic name for
the input TCRs in the plots (e.g., Epitope Specific). Avoid names with more
than 20 characters}

\item{input2.name}{(default="Reference"): If a second set of TCRs is provided
(i.e., input2 != ""), Provide a generic name for the input2 TCRs in the
plots. Avoid names with more than 20 characters.}

\item{output.format}{(default="pdf"): Choose the format for the output plots.}
}
\value{
Nothing.
}
\description{
!!! Give some additional description of the function.
}
