% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotVJ}
\alias{plotVJ}
\title{Plot the comparison between V or J usage.}
\usage{
plotVJ(
  count.es,
  count.rep,
  info,
  comp.baseline,
  pType = 1.2,
  sp = "HomoSapiens",
  ret.resList = F,
  combined.resList = NULL
)
}
\arguments{
\item{pType}{Switch telling the type of plot to use. See plot.VJ.switch
option from MixTCRviz.}

\item{sp}{Tells the species, which is used for gene colors in the plots.}

\item{ret.resList}{when T indicates to return a list of the results instead of
the plots directly (in this case, the "info" should have a 4th element,
indicating the model name).}

\item{combined.resList}{When this isn't NULL, we'll use the results from
this list to plot the results (from multiple models combined together).}
}
\description{
We should have different criteria for selecting the labels for the ES versus Pred.
}
