% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_model_colPals}
\alias{set_model_colPals}
\title{Defines the color palette for plots with models showed together.}
\usage{
set_model_colPals(models)
}
\arguments{
\item{models}{A vector of the models that are currently used in the figure.}
}
\value{
A named vector of the colors to use for each model.
}
\description{
Little function that defines the color palette to use for the different
models that are combined together when plot.modelsCombined is T.
These are the colors used around the bars and for the length distribution plots.
}
