% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_function.R
\name{clean.name.allele}
\alias{clean.name.allele}
\title{Correcting gene / allele names.}
\usage{
clean.name.allele(g, a, sp)
}
\description{
Take the gene + allele.
If allele, try to correct the gene*allele and return it.
If not possible, try to correct only the gene, and return only the gene
If no allele, try to correct the gene if needed
Cases that could not be corrected are returned as NA
}
