% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRviz.R
\name{MixTCRviz}
\alias{MixTCRviz}
\title{MixTCRviz: Plot TCR motifs}
\usage{
MixTCRviz(
  input1,
  output.path,
  input2 = NULL,
  baseline = NULL,
  use.allele = F,
  correct.gene.names = T,
  use.mouse.strain = F,
  check.cdr3.mode = 1,
  renormVJ = NULL,
  N.min = 1,
  output.stat = F,
  output.processed.data = F,
  set.cdr3a.length = NA,
  set.cdr3b.length = NA,
  species.default = "HomoSapiens",
  model.default = "Model_default",
  verbose = 1,
  plot = T,
  plot.cdr12.motif = F,
  plot.oneline = 0,
  plot.all.length = F,
  plot.cdr3.norm = 0,
  plot.VJ.switch = 1,
  plot.modelsCombined = FALSE,
  label.neg = F,
  label.min.fr = c(0.05, 0.05),
  chain.list.output = "AB",
  input1.name = "Input",
  input2.name = NULL,
  output.format = "pdf",
  print.size = T
)
}
\arguments{
\item{input1}{.csv files or data.frame with the input TCRs.
\itemize{
\item Columns should include "TRAV","TRAJ","cdr3_TRA","TRBV","TRBJ","cdr3_TRB" (one chain is allowed is chain.list.output = "A" or "B", respectively)
\item If TCRs from multiple eperiments/epitopes/classses/... are provided in the same file, the "model" column should indicate the models
("Model_default" is used by default if no "model" column is provided and model.default=NULL, see below). The use of multiple model is only possible in input1.
\item If TCRs from multiple species are provided in the same file, the "species" should indicate the species of the TCRs
("HomoSapiens" or "MusMusculus", with "HomoSapiens" being the default if no "species" column and species.default=NULL)
\item The "TRAV", "TRAJ", "TRBV", "TRBJ" should follow the IMGT
nomenclature, with or without allele (see below for potential name correction).
If a column is missing, empty values will be used.
\item The "cdr3_TRA" and "cdr3_TRB" columns should provide CDR3A/CDR3B sequences, following the standard definition (e.g., CAVNSDGQKLLF).
Cases with non-amino acid characters, or length < 7 or > 22 will be not be considered (i.e., put to NA).
If a column is missing, empty values will be used.
\item Other than .csv filename or data.frame, a third alternative consists of a list or .rds file with precomputed statistics, such as those generated by MixTCRviz in output.path/stats/.rds files.
This last option assumes all TCR come from the same model.
}}

\item{output.path}{name of the output directory (if not already existing, it
will be created). If existing the files with the same name will be overwritten.}

\item{input2}{(default=NULL): .csv file or data.frame containing a second set of
TCRs to be used in comparisons. Same format as input1, except that all TCRs are assumed to come from a single model so any data in the "model" field is omitted.
If input2 is provided, the comparisons is performed with this second input, and not the baseline
repertoire. In this case, all models provided in input1 will be compared with the repertoire in input2, and multiple species should not be mixed in input1.
Other than .csv filename or data.frame, a third alternative consists of a list or .rds with precomputed statistics,
such as those generated by MixTCRviz in output.path/stats/.rds files.}

\item{baseline}{(default=NULL) R object (list) or .rds/.rda/.rdata file containing all data about the
baseline repertoire to be used, in the same format as the default repertoires.
If .rda file, the object containing the data about the repertoires must be called 'baseline'.
If empty, the default baseline repertoires are used.}

\item{use.allele}{(default=FALSE)
\itemize{
\item FALSE: All V/J alleles are merged at the gene level (recommended).
\item TRUE: Alleles are kept. If some entries do not include alleles, the most frequent one is added.
Currently, use.allele is fixed to FALSE for mouse TCRs.
}}

\item{correct.gene.names}{(default=TRUE)
\itemize{
\item F: Do not attempt to correct V/J gene names. Put to NA genes not in IMGT.
\item T: Attempt to correct V/J gene names not in IMGT based on our internal dictionary.
Put to NA genes that could not be corrected
}}

\item{use.mouse.strain}{(default=FALSE)
\itemize{
\item FALSE: Merge the different TRAV genes corresponding to different mouse strains (e.g., TRAV10D, TRAVN10 and TRAV10 become all TRAV10).
This is recommended since differences between input TCRs and baseline TCRs
may result from the use of different mouse strains, and not be related to
specificity in input TCRs. In addition different TCR reconstruction tools
use different approaches to call these genes, which can create artificial
enrichment in one of them versus the baseline.
\item TRUE: The different TRAV segments (e.g., TRAV10D, TRAVN10 and TRAV10) are treated separately.
}}

\item{check.cdr3.mode}{(default=1)
\itemize{
\item 0: Keep all CDR3 without any correction.
\item 1: Remove V and CDR3 when the first CDR3 amino acid is incompatible with the V segment;
Remove J and CDR3 when the last two amino acids are not compatible with the J segments.
}}

\item{renormVJ}{(default=NULL)
\itemize{
\item NULL: If left empty, 1 is used if comparison is performed with a baseline (input2==NULL)
and 0 is used if comparison is performed with input2 (i.e., input2 != NULL)
\item 0: Compare CDR3 length distribution and motif with those from the baseline repertoire or input2
\item 1: Compare CDR3 length distribution and motif with those from the baseline repertoire with the V-J
usage observed in the input TCRs. In the plots the mark " | P(VJ)" is used to
indicate that CDR3 length distributions and motifs correspond to those
expected with the V-J usage in the TCRs used for training.
}}

\item{N.min}{(default=1) Minimum number of TCR (i.e., V-J-CDR3) for at least
one chain. This number is computed after cleaning the data.}

\item{output.stat}{(default=FALSE) If T, Create a stat/ folder with .rds objects summarizing the raw statistics for each model.
This includes countL, countV, countJ, countCDR3.L, etc. for each chain used in input.}

\item{output.processed.data}{(default=FALSE) If T, Create a processed_data/ folder with the data for each model after the different processing steps
(e.g., removing alleles, correcting V/J names, removing inconsistent CDR3, etc. depending on the cleaning option selected).}

\item{set.cdr3a.length}{(default=NA) Length for the CDR3a motif to be shown in the main plot.
By default the value corresponding to the most frequent CDR3a length in input1 (and also present in input2 if input2 is given) is chosen.}

\item{set.cdr3b.length}{(default=NA) Length for the CDR3b motif to be shown in the main plot.
By default the value corresponding to the most frequent CDR3b length in input1 (and also present in input2 if input2 is given) is chosen.}

\item{species.default}{(default="HomoSapiens"). Option to provide the species for all the TCR in input.
This is useful if your input does not contain a "species" column.
In case the input contains the "species" column, species.default is not considered.
Should be either "HomoSapiens" or "MusMusculus"}

\item{model.default}{(default="Model_default") Option to provide the model for all the TCR in input, which is also used as the name of the output files.
This is useful if your input does not contain a "model" column.
In case the input contains the "model" column, model.default is not considered.}

\item{verbose}{(default=1)
\itemize{
\item 0: Do not write any QC in the output
\item 1: Write max 10 examples of putative issues with the data (V/J names, CDR3 sequences, etc.) in the terminal
\item 2: Write all the putative issues with the data (V/J names, CDR3 sequences, etc.) in the terminal
\item 3: Write all putative issues + cases that were corrected
}}

\item{plot}{(default=TRUE)
\itemize{
\item F: Do not write the pdf/png/jpg files with the motifs
\item T: Write the pdf/png/jpg files with the motifs in output.path/
}}

\item{plot.cdr12.motif}{(default=FALSE)
\itemize{
\item F: Only show motifs for CDR3 of the most frequent length.
\item T: Include sequence motifs for CDR1 and CDR2. With this choice, plot.oneline is set to 0.
}}

\item{plot.oneline}{(default=0)
\itemize{
\item 0: Show the data on two lines (better for clarity).
\item 1: Show all plots in a single line (can be useful to compare different models).
\item 2: Show only V/J usage and length (i.e., do not show CDR3 motifs for the most frequent CDR3 length)
}}

\item{plot.all.length}{(default=FALSE)
\itemize{
\item F: Show only the CDR3 motifs for the most frequent CDR3 length.
\item T: Write in output.path/CDR3_length separate plot the V usage, J usage and CDR3 motifs for all CDR3 length, for both alpha and beta chains.
}}

\item{plot.cdr3.norm}{(default=0)
\itemize{
\item 0: Show the CDR3 motifs of the input and of the baseline repertoire or input2, possibly | P(VJ).
\item 1: Show the CDR3 motifs of the input TCRs after subtracting the baseline repertoire (not recommended).
\item 2: Show the CDR3 motifs of the input TCRs after normalising by the baseline repertoire (motif of normalised fold-change, not recommended)
}}

\item{plot.VJ.switch}{(default=1)
\itemize{
\item 1: Show the VJ usage as a scatter plot with inner colors, outer colors and shapes of the points based
on the V/J gene names (see lookup table in MixTCRviz/figures/, scheme1).
\item 1.2: Show the VJ usage as a scatter plot with colors of the points based
on the V/J gene names (see lookup table in MixTCRviz/figures/, scheme2).
\item 1.3: Show the VJ usage as a scatter plot with black points and V/J gene label in colors (see lookup table in MixTCRviz/figures/, scheme2).
\item 2: Show the VJ usage as a bar plot (see lookup table in MixTCRviz/figures/, scheme2).
}}

\item{plot.modelsCombined}{(default=FALSE)
\itemize{
\item FALSE or empty string: Show the data for each model separately.
\item TRUE or a string: Show the data for all models combined in a single
figure (based on the format plot.VJ.switch=2). When given as a
logical, the default name "modCombined" is used. Otherwise, the string
is used for the resulting figure filename.
}}

\item{label.neg}{(default=FALSE): If TRUE, show also the labels of the genes most depleted in input1.}

\item{label.min.fr}{(default=c(0.05,0.05)): Region (rectangle) of the left corner of V/J plots with no gene label.}

\item{chain.list.output}{(default="AB")
\itemize{
\item A: Only the alpha chain is plotted in output;
\item B: only the beta chain is plotted in output;
\item AB both chains are plotted in output
}}

\item{input1.name}{(default="Input"): Provide a generic name for
the input TCRs in the plots (e.g., Epitope Specific). Avoid names with more
than 20 characters}

\item{input2.name}{(default=NULL): If a second set of TCRs is provided
(i.e., input2 != NULL), Provide a generic name for the input2 TCRs in the
plots. Avoid names with more than 20 characters.}

\item{output.format}{(default="pdf"): Choose the format for the output
plots (can be "pdf", "png" or "jpg").}

\item{print.size}{(default=TRUE): If TRUE, print the number of TCRs in input1 in the plots.}
}
\value{
Nothing.
}
\description{
MixTCRviz is an R package to display TCR motifs for a set of TCRs provided by the user.
Typically, the input TCRs correspond to TCRs binding to a specific epitope, or isolated in a specific experiment.
MixTCRviz compares V usage, J usage, CDR3 length distribution and CDR3 sequence motifs to those expected from baseline TCR repertoire.
}
