% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixTCRviz.R
\name{MixTCRviz}
\alias{MixTCRviz}
\title{MixTCRviz: Plot TCR motifs}
\usage{
MixTCRviz(input1="test/test.csv", output.path = YOUR_PATH, ...)
}
\arguments{
\item{input1}{Either a .csv or .txt or .tsv file with the input TCRs, or a data.frame with the input TCRs, or list in the MixTCRviz format.
\itemize{
\item Columns should ideally include 'TRAV','TRAJ','cdr3_TRA','TRBV','TRBJ','cdr3_TRB' (single chain are also allowed)
\item If TCRs from multiple experiments/epitopes/classses/... are provided in the same file, the 'model' column should indicate the models
("Model_default" is used by default if no 'model' column is provided and \code{model.default}=NULL, see below).
\item If TCRs from multiple species are provided in the same file, the 'species' column should indicate the species of the TCRs
("HomoSapiens" or "MusMusculus", with "HomoSapiens" being the default if no 'species' column and \code{species.default}=NULL).
\item The 'TRAV', 'TRAJ', 'TRBV', 'TRBJ' should follow the IMGT nomenclature, with or without allele (see below for V/J name correction options).
\item The 'cdr3_TRA' and 'cdr3_TRB' columns should provide CDR3A/CDR3B sequences, following the standard definition (e.g., CAVNSDGQKLLF).
Cases with non-amino acid characters, or length < 7 or > 22 will be not be considered. See below for CDR3 sequence QC.
\item Other column names and/or input format are supported (see README.md).
}}

\item{output.path}{Name of the output directory (default=NULL). If not already existing, it
will be created. If existing, files with the same name will be overwritten.
If left empty, nothing will be written, and only R objects will be returned.}

\item{input2}{.csv, .tsv of tab delimited .txt filename, or data.frame containing a second set of
TCRs to be used in comparisons. Same format as input1, except that all TCRs are assumed to come from a single model so any data in the 'model' column is omitted.
If input2 is provided, the comparisons is performed with this second input, and not the baseline
repertoire. In this case, all models provided in input1 will be compared with the repertoire in input2, and multiple species should not be mixed in input1.
Other than filename or data.frame, a third alternative consists of a list or .rds file with precomputed statistics,
such as those generated by MixTCRviz::build_stat() function. default=NULL.}

\item{baseline}{R object (list) or name of a .rds/.rda/.rdata file containing all data about the
baseline repertoire to be used, in the same format as the default repertoires.
If .rda file, the object containing the data about the repertoires must be called 'baseline'.
Alternatively one of the strings: "Default" or "SEQTR" for data in input1 generated by publicly available TCR-seq approaches (Default) or SEQTR (SEQTR) protocols.
If empty, the default baseline repertoires are used. default=NULL.}

\item{chain}{Decide which chain to show in the motifs.
\itemize{
\item "AB" (default): both chains are plotted in output.
\item "A": Only the alpha chain is plotted in output.
\item "B": Only the beta chain is plotted in output.
}}

\item{interactive.plots}{Decide whether to plot interactive html plots.
\itemize{
\item FALSE (default): Do not create an html file with interactive plots.
\item TRUE: Create an html file with interactive plots. Only applicable if \code{output.path} is set and \code{plot}=T. Takes slightly longer
}}

\item{use.allele}{Decide whether alleles should be used
\itemize{
\item FALSE (default): All V/J alleles are merged at the gene level (recommended).
\item TRUE: Alleles are kept. If some entries do not include alleles, the most frequent one is added.
Currently, \code{use.allele} is fixed to FALSE for mouse TCRs.
}}

\item{correct.gene.names}{Decide how to correct V/J names
\itemize{
\item FALSE: Do not attempt to correct V/J gene names. Put to NA genes not in IMGT.
\item TRUE (default): Attempt to correct V/J gene names not in IMGT based on our internal dictionary.
V/J names which could not be corrected are put to NA.
}}

\item{use.mouse.strain}{Decide whether or not to merge TRAV genes corresponding to different strains (e.g., TRAV10, TRAV10D, TRAVN10)
\itemize{
\item FALSE (default): Merge the different TRAV genes corresponding to different mouse strains (e.g., TRAV10D, TRAVN10 and TRAV10 become all TRAV10).
This is recommended the recommended option.
\item TRUE: The different TRAV segments (e.g., TRAV10D, TRAVN10 and TRAV10) are treated separately.
In general, this is NOT recommended since the baseline is challenging to establish. Should be used only if you have a baseline coming from same mice used in input1.
}}

\item{check.cdr3.mode}{Decide how to check CDR3 sequences for compatibility with V/J genes
\itemize{
\item 0: Keep all CDR3 without any correction.
\item 1 (default): Remove V and CDR3 when the first CDR3 amino acid is incompatible with the V segment;
Remove J and CDR3 when the last 2 amino acids are not compatible with the J segments.
Allow compatibility check with any allele.
}}

\item{renormVJ}{Decide whether or not to build baseline length distribution and CDR3 motif for P(VJ) in input1.
\itemize{
\item NULL: If left empty, 1 is used if comparison is performed with a baseline (\code{input2}==NULL)
and 0 is used if comparison is performed with input2 (i.e., \code{input2} != NULL)
\item 0: Compare CDR3 length distribution and motif with those from the baseline repertoire or input2
\item 1 (default): Compare CDR3 length distribution and motif with those from the baseline repertoire with the V-J
usage observed in the input TCRs. In the plots the mark " | P(VJ)" is used to
indicate that CDR3 length distributions and motifs correspond to those
expected with the V-J usage in the TCRs used for training.
}}

\item{N.min}{Minimum number of TCRs (i.e., V-J-CDR3) for at least
one chain. This number is computed after cleaning the data (default=1) .}

\item{output.stat}{Print the stats in a file.
\itemize{
\item FALSE (default): Do not print the stats
\item TRUE: create a output.path/stat/ folder with .rds files summarizing the raw statistics for each model. Requires \code{output.path} to be initialized.
}}

\item{output.processed.data}{Print the processed data (after the different cleaning steps).
\itemize{
\item FALSE (default): Do not print the processed/clean data to a file
\item TRUE: create a output.path/processed_data/ folder with the data for each model after the different processing steps
(e.g., removing alleles, correcting V/J names, removing inconsistent CDR3, etc. depending on the cleaning option selected). Requires \code{output.path} to be initialized.
}}

\item{filename.output}{Provide a name for the output file (default=NULL).
This option should not be used if input1 contains data from multiple models.
If NULL, the value(s) in the 'model' column of input1 are used (i.e., input1$model).
If no 'model' column is provided, the value in \code{model.default} is used.}

\item{set.cdr3a.length, set.cdr3a.length}{Length for the CDR3 motif to be shown in the main plot (default=NA).
If left empty, the value corresponding to the most frequent CDR3 length in input1 (and also present in input2 if input2 is provided) is chosen.}

\item{plot.title}{If TRUE (default), print the model name as title to the plots.}

\item{set.title}{Set the title of the plots (default=NULL).
If NULL, input1$model are used as titles. If no 'model' column is provided in input1, \code{model.default} is used.}

\item{species.default}{Option to provide the species for all the TCR in input1 (default="HomoSapiens").
This is useful if your input does not contain a 'species' column.
In case the input contains the 'species' column, \code{species.default} is not considered.
Should be either "HomoSapiens" or "MusMusculus"}

\item{model.default}{Option to provide the model for all the TCR in input1 (default="Model_default"), which is also used as the name of the output files.
This is useful if your input does not contain a 'model' column.
In case the input contains the 'model' column, \code{model.default} is not considered.}

\item{verbose}{Decide how much information is given in the R Console.
\itemize{
\item 0: Do not write any QC in the output
\item 1 (default): Write max 10 examples of putative issues with the data (V/J names, CDR3 sequences, etc.) in the Console
\item 2: Write all the putative issues with the data (V/J names, CDR3 sequences, etc.) in the Console
\item 3: Write all putative issues + cases that were corrected (can be long depending on your data).
}}

\item{build.clones}{Decide whether to reconstruct the clones if the data are provided in format with clone.id (default=FALSE).
If TRUE and if the data are provided in format with clone.id/complex.id/..., reconstruct the actual clones in output}

\item{plot}{Decide whether or not to plot the motifs.
\itemize{
\item FALSE: Do not build the motifs. This overrides \code{plot.all.length}=T and \code{interactive.plots}=T
\item TRUE (default): Build the motifs
}}

\item{plot.cdr12.motif}{Decide whether or not to show the motifs for CDR1 and CDR2.
\itemize{
\item FALSE (default): Only show motifs for CDR3.
\item TRUE: Include sequence motifs for CDR1 and CDR2. With this choice, \code{plot.oneline} is set to 0. This information is redundant with V usage.
}}

\item{plot.oneline}{Decide how to plot the motifs.
\itemize{
\item 0 (default): Show the data on two lines (better for clarity).
\item 1: Show all plots in a single line (can be useful to compare different models).
\item 2: Show only V/J usage and length (i.e., do not show CDR3 motifs).
}}

\item{plot.all.length}{Decide whether to plot the motifs for each CDR3 length.
\itemize{
\item FALSE (default): Show only the CDR3 motifs for the most frequent CDR3 length.
\item TRUE: Write in \code{output.path}/CDR3_length distinct plots for the V usage, J usage and CDR3 motifs
for all CDR3 lengths. Only applicable if \code{plot}=T.
}}

\item{plot.cdr3.norm}{Decide how to build the motifs for the baseline.
\itemize{
\item 0 (default): Show the CDR3 motifs of the input and of the baseline repertoire or input2, possibly | P(VJ).
\item 1: Show the CDR3 motifs of the input TCRs after subtracting the baseline repertoire (not recommended due to poor readability).
\item 2: Show the CDR3 motifs of the input TCRs after normalising by the baseline repertoire (motif of normalised fold-change, not recommended due to poor readability)
}}

\item{plot.VJ.switch}{Change the color and/or format of the motifs.
\itemize{
\item 1 (default): Show the VJ usage as a scatter plot with inner colors, outer colors and shapes of the points based
on the V/J gene names (see lookup table in MixTCRviz/figures/, scheme1).
\item 1.2: Show the VJ usage as a scatter plot with colors of the points based
on the V/J gene names (see lookup table in MixTCRviz/figures/, scheme2).
\item 1.3: Show the VJ usage as a scatter plot with black points and V/J gene label in colors (see lookup table in MixTCRviz/figures/, scheme2).
\item 2: Show the VJ usage as a bar plot (see lookup table in MixTCRviz/figures/, scheme2).
}}

\item{plot.modelsCombined}{Decide whether to combine the different models.
\itemize{
\item FALSE (default) or empty string: Show the data for each model separately.
\item TRUE or a string: Show the data for all models combined in a single
figure (based on the format \code{plot.VJ.switch}=2). When given as a
logical, the default name "modCombined" is used. Otherwise, the string
is used for the resulting figure filename.
}}

\item{label.neg}{Show the label of V/J genes negatively enriched in input1 (default=FALSE).
If TRUE, show also the labels of the genes most depleted in input1.}

\item{label.diag}{Decide to keep some label in the upper right corner, based on label.diag value (default=0.3).
This can be especially useful when comparing two epitope-specific TCR repertoires.}

\item{plot.sd}{Decide whether to include standard deviation in P(V), P(J) and P(L) plots.
\itemize{
\item TRUE (default): Show standard deviation for P(V), P(J) and P(L), if such data are provided for baseline and/or input1.
\item FALSE: Do not show standard deviation
}}

\item{label.min.fr.x, label.min.fr.y}{Region (i.e., X - Y rectangle) of the left corner of V/J plots with no gene label (default=0.05).}

\item{keep.incomplete.chain}{Decide whether to keep incomplete alpha or beta chains.
\itemize{
\item TRUE (default): Incomplete chains are kept.
\item FALSE: Incomplete chains are discarded. Even if input1 only consists of complete chains, incomplete chains can occur when one V/J gene cannot be corrected,
or when there is some incompatibilities between V/J names and CDR3 sequences
}}

\item{input1.name, input2.name}{Provide a generic name for the TCRs in input1 or input2 (default="Input" and "Baseline"/"Input2"). Avoid names with more
than 20 characters.}

\item{output.format}{Choose the format for the output, can be "pdf" (default), "png" or "jpg".}

\item{keep.colnames.origin}{Decide whether to keep the original names of the columns (default=FALSE).
If TRUE, keep the name of the columns dedicated the TCRs in the input. The option is not supported for format with clone.id}

\item{print.size}{If TRUE (default), print the number of TCRs in input1 in the plots.}
}
\value{
R list containing the plots (if \code{plot}=TRUE), the processed data (if input1 is a file or data.frame) and an R list with the stats for each model in input1.
}
\description{
MixTCRviz is an R package to display TCR binding motifs for a set of TCRs provided by the user.
Typically, the input TCRs correspond to TCRs binding to a specific epitope, or isolated in a specific experiment.
MixTCRviz compares V usage, J usage, CDR3 length distribution and CDR3 sequence motifs to those expected from baseline TCR repertoire.
}
