% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotVJ}
\alias{plotVJ}
\title{Plot the comparison between V or J usage.}
\usage{
plotVJ(
  count.es,
  count.rep,
  sd.es = NULL,
  sd.rep = NULL,
  info,
  comp.baseline,
  pType = 1,
  species = "HomoSapiens",
  ret.resList = F,
  combined.resList = NULL,
  label.neg = F,
  label.diag = 0.3,
  label.min.fr = c(0.05, 0.05),
  print.size = T,
  plot.sd = T,
  verbose = 1
)
}
\arguments{
\item{pType}{Switch telling the type of plot to use. See plot.VJ.switch
option from MixTCRviz.}

\item{species}{Tells the species, which is used for gene colors in the plots.}

\item{ret.resList}{when T indicates to return a list of the results instead of
the plots directly (in this case, the "info" should have a 4th element,
indicating the model name).}

\item{combined.resList}{When this isn't NULL, we'll use the results from
this list to plot the results (from multiple models combined together).}

\item{label.neg}{(default=F): If T, show also the labels of the genes most depleted in input1}

\item{label.diag}{(default=0.3): Print label on the diagonal above a certain value for both x and y axis}

\item{label.min.fr}{(default=c(0.05, 0.05)): Region (rectangle) of the left corner of V/J plots with no gene label}
}
\description{
We should have different criteria for selecting the labels for the ES versus Pred.
}
