% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotVJ}
\alias{plotVJ}
\title{Plot the comparison between V or J usage.}
\usage{
plotVJ(
  count.es,
  count.rep,
  info,
  comp.baseline,
  as.bars = F,
  sp = "HomoSapiens"
)
}
\arguments{
\item{as.bars}{When T, we plot the enrichment as bar plot instead of scatter plot.}

\item{sp}{Tells the species, which is used for bar colors when as.bars=T.}
}
\description{
We should have different criteria for selecting the labels for the ES versus Pred.
}
